+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-# Changelog
+## 0.1.1 (2022-04-XX)
 
-## [0.1.0] - 2022-03-29
+* Switched to asyncio
+* Added profile management to Board class
+* Added tests for Board and Profile classes
+* Switched to dataclasses for message struct generation
+
+## 0.1.0 (2022-03-29)
 
 * First release on PyPI.